+++ utils/style_loader.py
@@ -1,5 +1,12 @@
 # utils/style_loader.py
 
+"""
+Stylesheet loading utilities.
+
+Provides helper functions for loading and applying Qt stylesheets 
+(.qss) coupled with the theme stylesheets. 
+"""
+
 from utils.util import resource_path
 
 from PyQt6.QtWidgets import QApplication
@@ -7,11 +14,27 @@ from PyQt6.QtWidgets import QApplication
 def load_stylesheet(
         app: QApplication, 
         theme: str = "dark", 
-        base_path: str = "assets/styles/styles.qss" 
-        # themes_path: str = "assets/themes"
+        base_path: str = "assets/styles/styles.qss",
+        themes_path: str = "assets/themes/"
     ) -> None:
+
+    """
+    Loads and applies application stylesheets.
+
+    Combines the shared base stylesheet (by default `styles.qss`) with
+    the specified theme stylesheet and applies the result globally to
+    the QApplication instance.
+
+    Any file loading errors will be caught and printed to the console.
     
-    theme_path = f"assets/themes/{theme}.qss"
+    Args:
+        app: The instance of QApplication to apply the styles to
+        theme: Theme name used to select the theme stylesheet
+        base_path: The path to the base stylesheet file
+        themes_path: The path to the themes directory
+    """
+    
+    theme_path = f"{themes_path}/{theme}.qss"
 
     try: 
         with open(resource_path(base_path), "r") as f: